customize Bash Prompt for Network Namespace in Mininet
Using a different prompt for a bash terminal when using network
namespace serves the purpose of clearly indicating that the user is
running commands within a specific node in a network setup.

By customizing the prompt with the hostname and directory path, it helps
differentiate the context in which the user is operating. This can be
particularly useful in Mininet, where multiple network nodes may be
running concurrently, allowing users to easily distinguish which node
they are currently interacting with.

For example, assuming that a Mininet network includes a node named h1,
after logging in to the h1 terminal, the prompt will look like the
following:

h1:~/itp-labs>

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>